Revert "[Member] 회원가입 휴대폰 인증 완료"
This reverts commit 113dad10d225965e063b137467e353612d411e30.

+++ src/main/java/com/eob/shop/model/data/ShopEntity.java
@@ -175,10 +175,10 @@ public class ShopEntity {
     /**
      * 상점 위치 - 경도
      */
-    private Double longitude;
+    private float longitude;
 
     /**
      * 상점 위치 - 위도
      */
-    private Double latitude;
+    private float latitude;
 }
